Refactor: move to module project
