+++ Source/Infrastructure/Iwentys.Infrastructure.Application/Iwentys.Infrastructure.Application.csproj
@@ -20,4 +20,8 @@
     <ProjectReference Include="..\Iwentys.Infrastructure.Configuration\Iwentys.Infrastructure.Configuration.csproj" />
     <ProjectReference Include="..\Iwentys.Infrastructure.DataAccess\Iwentys.Infrastructure.DataAccess.csproj" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Modules\AccountManagment" />
+  </ItemGroup>
 </Project>
