--- Source/Infrastructure/Iwentys.Infrastructure.Application/Modules/AccountManagment/MentorsManagment/Commands/AddMentor.cs
@@ -1,7 +1,0 @@
-﻿namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.MentorsManagment.Commands
-{
-    public class AddMentor
-    {
-        
-    }
-}
