--- Source/Infrastructure/Iwentys.Infrastructure.Application/Modules/AccountManagment/MentorsManagment/Commands/RemoveMentorFromGroup.cs
@@ -1,7 +1,0 @@
-﻿namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.MentorsManagment.Commands
-{
-    public class RemoveMentorFromGroup
-    {
-        
-    }
-}
