--- Source/Infrastructure/Iwentys.Infrastructure.Application/Modules/AccountManagment/MentorsManagment/Queries/GetMentorsByGroupId.cs
@@ -1,7 +1,0 @@
-﻿namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.MentorsManagment.Queries
-{
-    public class GetMentorsByGroupSubjectId
-    {
-        
-    }
-}
