+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/Commands/RemoveMentorFromGroup.cs
@@ -1,0 +1,7 @@
+﻿namespace Iwentys.Modules.AccountManagement.Mentors.Commands
+{
+    public class RemoveMentorFromGroup
+    {
+        
+    }
+}
