+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/Dtos/GroupMentorsDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.Dtos
+namespace Iwentys.Modules.AccountManagement.Mentors.Dtos
 {
     public class GroupMentorsDto
     {
