+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/MentorsManagementController.cs
@@ -1,17 +1,18 @@
 ﻿using System.Threading.Tasks;
-using Iwentys.Infrastructure.Application.Modules.AccountManagment.MentorsManagment.Queries;
+using Iwentys.Infrastructure.Application;
+using Iwentys.Modules.AccountManagement.Mentors.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Iwentys.Infrastructure.Application.Modules.AccountManagment.MentorsManagment
+namespace Iwentys.Modules.AccountManagement.Mentors
 {
     [Route("api/account-management/mentors")]
     [ApiController]
-    public class MentorsManagmentController : ControllerBase
+    public class MentorsManagementController : ControllerBase
     {
         private readonly IMediator _mediator;
 
-        public MentorsManagmentController(IMediator mediator)
+        public MentorsManagementController(IMediator mediator)
         {
             _mediator = mediator;
         }
