+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/Queries/GetMentorsByGroupId.cs
@@ -1,0 +1,7 @@
+﻿namespace Iwentys.Modules.AccountManagement.Mentors.Queries
+{
+    public class GetMentorsByGroupSubjectId
+    {
+        
+    }
+}
